Add global progress tracking system and level-up notifications
Implement a centralized progress context to manage user points, levels, and streaks across the application. Refactor components to utilize the global context for point updates, ensuring level-up and streak milestone popups are displayed consistently on all screens. Update navigation structure and add loading and welcome screens for improved user experience.
