+++ LEVEL_UP_FIX.md
@@ -1,0 +1,189 @@
+# Level-Up Popup Fix - Complete Resolution
+
+## Problem Identified
+
+The level-up popup was not appearing on quiz pages, flashcard screens, or other pages because:
+
+1. **Stale State Issue**: The `incrementPoints` function in `ProgressContext` was comparing `newLevel` against `stats.level` from React state, which could be outdated due to asynchronous state updates.
+
+2. **Direct AsyncStorage Writes**: Quiz pages and flashcard screen were directly writing to AsyncStorage instead of using the global `incrementPoints` function, completely bypassing the level-up detection logic.
+
+## Fixes Applied
+
+### 1. Fixed Progress Context (`app/contexts/ProgressContext.tsx`)
+
+**Before:**
+```typescript
+const incrementPoints = useCallback(async (points: number) => {
+  const currentPoints = stats.points + points;  // ❌ Using stale state
+  const newLevel = Math.floor(currentPoints / 100) + 1;
+  
+  if (newLevel > stats.level) {  // ❌ Comparing against stale state
+    setLevelUpModalVisible(true);
+  }
+}, [stats.level, stats.points]);  // ❌ Recreated on every state change
+```
+
+**After:**
+```typescript
+const incrementPoints = useCallback(async (points: number) => {
+  // ✅ Read from AsyncStorage for accuracy
+  const storedPoints = await AsyncStorage.getItem('@VT_POINTS');
+  const currentTotalPoints = storedPoints ? parseInt(storedPoints) : 0;
+  const currentLevel = Math.floor(currentTotalPoints / 100) + 1;
+  
+  const newTotalPoints = currentTotalPoints + points;
+  const newLevel = Math.floor(newTotalPoints / 100) + 1;
+  
+  // ✅ Compare against current level from storage, not state
+  if (newLevel > currentLevel) {
+    console.log(`🎉 LEVEL UP! ${currentLevel} → ${newLevel}`);
+    setLevelUpModalVisible(true);
+  }
+}, [incrementStreak]);  // ✅ Stable dependency
+```
+
+### 2. Fixed Quiz Pages
+
+**Files Updated:**
+- `app/quiz/easy.tsx`
+- `app/quiz/medium.tsx`
+- `app/quiz/hard.tsx`
+
+**Before:**
+```typescript
+// ❌ Direct AsyncStorage write bypasses level-up detection
+const currentPoints = await AsyncStorage.getItem('@VT_POINTS');
+const newTotal = (currentPoints ? parseInt(currentPoints) : 0) + points;
+await AsyncStorage.setItem('@VT_POINTS', newTotal.toString());
+```
+
+**After:**
+```typescript
+// ✅ Uses global context - triggers level-up popup automatically
+await incrementPoints(points);
+```
+
+### 3. Fixed Flashcard Screen
+
+**File Updated:**
+- `app/flash_card.tsx`
+
+**Before:**
+```typescript
+// ❌ Direct AsyncStorage write
+const currentPoints = await AsyncStorage.getItem('@VT_POINTS');
+const newTotal = (currentPoints ? parseInt(currentPoints) : 0) + bonusPoints;
+await AsyncStorage.setItem('@VT_POINTS', newTotal.toString());
+```
+
+**After:**
+```typescript
+// ✅ Uses global context
+await incrementPoints(bonusPoints);
+```
+
+## How It Works Now
+
+### Level-Up Detection Flow
+
+1. **User earns points** (quiz, learned word, review, etc.)
+2. **Component calls** `await incrementPoints(points)`
+3. **Context reads** current points from AsyncStorage (source of truth)
+4. **Calculates** new level: `Math.floor(newTotalPoints / 100) + 1`
+5. **Compares** new level vs current level
+6. **If leveled up** → Sets modal data and triggers modal visibility
+7. **Modal appears** instantly on current screen with celebration
+
+### Debug Logging
+
+Added helpful console logs to track points and level-ups:
+
+```
+💎 Points awarded: +10 (95 → 105, Level 1)
+🎉 LEVEL UP! 1 → 2 (105 points)
+```
+
+This makes it easy to verify the system is working correctly in development.
+
+## Points System Summary
+
+| Action | Points | Trigger Location |
+|--------|--------|-----------------|
+| Add to Review | +5 | Home, Word Details |
+| Mark as Learned | +10 | Home, Word Details, Flashcards |
+| Easy Quiz (correct) | +10 per word | Quiz Easy |
+| Medium Quiz (correct) | +15 per word | Quiz Medium |
+| Hard Quiz (correct) | +20 per word | Quiz Hard |
+
+**Level System:**
+- Level = `floor(totalPoints / 100) + 1`
+- Level 1: 0-99 points
+- Level 2: 100-199 points
+- Level 3: 200-299 points
+- etc.
+
+## Testing Checklist
+
+✅ **Quiz Pages**
+- Start quiz with 95 points (Level 1)
+- Answer 1 question correctly
+- Should see level-up popup appear on quiz screen
+
+✅ **Word Details**
+- Navigate to word details with 90 points
+- Mark as learned (+10 points)
+- Should see level-up popup appear on details screen
+
+✅ **Flashcards**
+- Start flashcards with 95 points
+- Mark a word as learned
+- Should see level-up popup appear on flashcard screen
+
+✅ **Home Screen**
+- Already working (verified)
+- Add to review or mark learned triggers popup
+
+✅ **Settings Screen**
+- No points awarded here, but popup should appear if triggered elsewhere
+
+## Code Quality Improvements
+
+1. **Single Source of Truth**: AsyncStorage is the source of truth for points
+2. **No State Race Conditions**: Don't rely on potentially stale React state
+3. **Consistent API**: All screens use same `incrementPoints` function
+4. **Better Debugging**: Console logs help track point flow
+5. **Cleaner Code**: Removed duplicate AsyncStorage logic
+
+## Expected Behavior
+
+When a user levels up:
+
+1. **Immediate Popup**: Appears on whatever screen they're on
+2. **Celebration**: Shows confetti, trophy icon, glowing badge
+3. **Stats Display**: Shows new level, total points, progress
+4. **Motivational Message**: Random encouraging message
+5. **Progress Bar**: Visual progress within current level
+6. **Tap to Close**: User can dismiss by tapping anywhere
+
+## Future Enhancements
+
+Potential improvements:
+- Sound effects on level-up
+- Animated confetti particles
+- Level-specific rewards or unlocks
+- Share level-up achievement
+- Level history tracking
+- Custom badges per level milestone
+
+## Verification
+
+All changes have been:
+- ✅ Implemented across all relevant files
+- ✅ Tested for linter errors (none found)
+- ✅ Documented with inline comments
+- ✅ Enhanced with debug logging
+- ✅ Verified for consistent API usage
+
+The level-up popup system now works **globally and consistently** across the entire app! 🎉
+
